# openapi.yaml
openapi: 3.0.1
info:
  title: FINEKO API
  version: 1.0.0
servers:
  - url: http://localhost:3000
paths:
...
  /companies/{companyId}/tasks:
    get:
      summary: Завдання компанії
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Масив завдань
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
    post:
      summary: Створити завдання
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
      responses:
        '201':
          description: Нове завдання
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
  /companies/{companyId}/tasks/{taskId}:
    patch:
      summary: Оновити завдання
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
      responses:
        '200':
          description: Оновлене завдання
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
    delete:
      summary: Видалити завдання
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Без вмісту
...
  /companies/{companyId}/org-structure:
    get:
      summary: Організаційна структура
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Масив вузлів
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrgNode'
    post:
      summary: Додати вузол
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrgNode'
      responses:
        '201':
          description: Новий вузол
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrgNode'
  /companies/{companyId}/org-structure/{nodeId}:
    patch:
      summary: Оновити вузол
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrgNode'
      responses:
        '200':
          description: Оновлений вузол
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrgNode'
    delete:
      summary: Видалити вузол
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Без вмісту
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...